% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chandwichGLM.R
\docType{package}
\name{chandwichGLM}
\alias{chandwichGLM}
\alias{chandwichGLM-package}
\title{chandwichGLM: Chandler-Bate Sandwich Loglikelihood Adjustment for Generalized
Linear Models}
\description{
Performs inferences from Generalized Linear Models fitted to cluster
correlated data, using adjustment to the independence loglikelihood
developed in
\href{http://dx.doi.org/10.1093/biomet/asm015}{Chandler and Bate (2007)}.
The adjustment are performed by the
\code{\link[chandwich]{chandwich}}.
}
\details{
Add details.

See Chandler and Bate (2007) for full details and
\code{vignette("chandwichGLM-vignette", package = "chandwich")} for an
overview of the package.
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
